hook up the mint transaction to the new state mgmt system
