+++ src/store/enum.ts
@@ -1,6 +1,7 @@
 export enum THUNK_PREFIX {
   'GET_OWNED_CRUCIBLED' = 'crucibles/getOwnedCrucibles',
   'TRANSFER_CRUCIBLE' = 'transactions/transferCrucible',
+  'MINT_CRUCIBLE' = 'transactions/mintCrucible',
 }
 
 export enum SLICE_NAME {
