+++ src/store/transactions/reducer.ts
@@ -1,22 +1,7 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { useWeb3React } from '@web3-react/core';
-import { useAppDispatch, useAppSelector } from '../hooks';
 import { SLICE_NAME } from '../enum';
-import {
-  TxnDetails,
-  TxnState,
-  TxnStatus,
-  TxnType,
-  UseTransactions,
-} from './types';
-import {
-  transferCrucible as _transferCrucible,
-  TransferCrucibleArgs,
-} from './actions/transferCrucible';
-import { useConfig } from '../config';
-import { useNotify } from '../../context/transactions';
-import { useToast } from '@chakra-ui/toast';
-import { useCrucibles } from '../../context/crucibles';
+import { TxnDetails, TxnState, TxnStatus } from './types';
+import { TransferCrucibleArgs } from './actions/transferCrucible';
 
 const updateLocalStorageSavedTransactions = (txn: TxnDetails) => {
   const savedTxns: TxnDetails[] = JSON.parse(
@@ -84,84 +69,5 @@ export const transactionsSlice = createSlice({
     },
   },
 });
-
-export const useTransactions = (): UseTransactions => {
-  const dispatch = useAppDispatch();
-  const { config } = useConfig();
-  const web3React = useWeb3React();
-  const { monitorTx } = useNotify();
-  const { reloadBalances, reloadCrucibles } = useCrucibles();
-  const configForNetwork = config[web3React.chainId || 1];
-  const { transactions } = useAppSelector((state) => state.transactions);
-  const toast = useToast();
-
-  const clearSavedTransactions = () => {
-    dispatch(transactionsSlice.actions.clearSavedTransactions());
-  };
-
-  const updateSavedTransaction = (updatedTx: Partial<TxnDetails>) =>
-    dispatch(
-      transactionsSlice.actions.setTransactionStatus({
-        ...updatedTx,
-        type: TxnType.transfer,
-        account: web3React.account as string,
-        chainId: web3React.chainId as number,
-      })
-    );
-
-  const transferCrucible = async (crucibleId: string, transferTo: string) => {
-    const transferAction = await dispatch(
-      _transferCrucible({
-        config: configForNetwork,
-        web3React,
-        monitorTx,
-        updateTx: updateSavedTransaction,
-        transferTo,
-        crucibleId,
-      })
-    );
-
-    // transfer success
-    if (_transferCrucible.fulfilled.match(transferAction)) {
-      reloadBalances();
-      reloadCrucibles();
-    }
-
-    // transfer failure
-    if (_transferCrucible.rejected.match(transferAction)) {
-      toast({
-        title: 'Transfer failed',
-        status: 'error',
-        duration: 2000,
-        isClosable: true,
-      });
-
-      updateSavedTransaction({
-        status: TxnStatus.Failed,
-      });
-    }
-  };
-
-  const pendingTransactions = transactions.filter(
-    (txn) => txn.status === TxnStatus.PendingOnChain
-  );
-
-  const completedTransactions = transactions.filter(
-    (txn) => txn.status === TxnStatus.Mined
-  );
-
-  const savedTransactions = transactions.filter(
-    (txn) =>
-      txn.account === web3React.account && txn.chainId === web3React.chainId
-  );
-
-  return {
-    savedTransactions,
-    pendingTransactions,
-    completedTransactions,
-    clearSavedTransactions,
-    transferCrucible,
-  };
-};
 
 export default transactionsSlice.reducer;
